feat: :art: Combine 2 profile pages to 1
