+++ frontend/src/hooks/index.ts
@@ -1,5 +1,6 @@
 export { useHandleSocket } from './useHandleSocket';
 export { useGetCurrentUser } from './useGetCurrentUser';
 export { useGetCertainUser } from './useGetCertainUser';
+// export { useGetUser } from './useGetUser';
 export { useGetCurrentChatRoom } from './useGetCurrentChatRoom';
 export { useLogout } from './useLogout';
