+++ frontend/src/hooks/useGetCertainUser.ts
@@ -2,7 +2,7 @@ import { useQuery } from 'react-query';
 import { useParams } from 'react-router-dom';
 import { useErrorHandler } from 'react-error-boundary';
 
-import { getCertainUserInfo } from 'services';
+import { getUserInfo } from 'services';
 import { UserInfoType } from 'types/user';
 
 const INIT_DATA: UserInfoType = {
@@ -18,7 +18,7 @@ const INIT_DATA: UserInfoType = {
 export function useGetCertainUser() {
   const handleError = useErrorHandler();
   const id = Number(useParams().id);
-  const { data = INIT_DATA } = useQuery([`User${id}`], () => getCertainUserInfo({ id }), {
+  const { data = INIT_DATA, refetch } = useQuery([`User${id}`], () => getUserInfo({ id }), {
     refetchOnWindowFocus: false,
     staleTime: 1000 * 60 * 60 * 12,
     cacheTime: 1000 * 60 * 60 * 12, // TODO: 적절한 시간으로 수정
@@ -27,6 +27,7 @@ export function useGetCertainUser() {
       handleError(e);
     },
   });
+  const isMine = isNaN(id);
 
-  return data;
+  return { data, isMine, refetch };
 }
