+++ frontend/src/services/index.ts
@@ -1,5 +1,5 @@
 export { getCurrentUserInfo } from './getCurrentUserInfo';
-export { getCertainUserInfo } from './getCertainUserInfo';
+export { getUserInfo } from './getUserInfo';
 export { getFtCallbackCode } from './getFtCallbackCode';
 export { putUserProfile } from './putUserProfile';
 export { deleteAuthSignout } from './deleteAuthSignout';
